[FIR] Move `RedundantCallOfConversionMethod` to the default checkers list
KT-77727

+++ compiler/testData/diagnostics/tests/smartCasts/objectLiterals/exclexcl.fir.kt
@@ -9,7 +9,7 @@ fun foo(): Int {
     if (c is Int) {
         val k = object: Runnable() {
             init {
-                a!!.toInt()
+                a!!
             }
             override fun run() = Unit
         }
